refactor: sort and rename component clients into dedicated dirs
