+++ crates/mempool_infra/src/component_client/mod.rs
@@ -1,0 +1,3 @@
+pub mod definitions;
+pub mod local_component_client;
+pub mod remote_component_client;
